chore(serve): split verify and sign in distinct files
