+++ pkg/cmd/serve/serve.go
@@ -131,100 +131,7 @@ func writeResponse(w http.ResponseWriter, r *types.Result) {
 	w.Write(b)
 }
 
-func signHander(state meta.Status, pubKey string) func(w http.ResponseWriter, r *http.Request) {
-	return func(w http.ResponseWriter, r *http.Request) {
-		s := state
-		p := pubKey
-		sign(s, p, w, r)
-	}
-}
-
-func sign(state meta.Status, pubKey string, w http.ResponseWriter, r *http.Request) {
-	decoder := json.NewDecoder(r.Body)
-	var jsonRequest api.ArtifactRequest
-	err := decoder.Decode(&jsonRequest)
-
-	if err != nil {
-		writeErrorResponse(w, "invalid request body", err, http.StatusBadRequest)
-		return
-	}
-
-	// check hash in request
-	if jsonRequest.Hash == "" {
-		writeErrorResponse(w, "hash cannot be empty", nil, http.StatusBadRequest)
-		return
-	}
-	if jsonRequest.Name == "" {
-		writeErrorResponse(w, "name cannot be empty", nil, http.StatusBadRequest)
-		return
-	}
-
-	user, err := currentUser()
-	if user == nil || err != nil {
-		writeErrorResponse(w, "no such user", err, http.StatusBadRequest)
-		return
-	}
-
-	if pubKey == "" {
-		pubKey = user.DefaultKey()
-	}
-	if pubKey == "" {
-		writeErrorResponse(w, "no key available", nil, http.StatusBadRequest)
-		return
-	}
-
-	// Make the artifact to be signed
-	var a api.Artifact
-	a.Hash = jsonRequest.Hash
-	a.Name = jsonRequest.Name
-	a.Size = jsonRequest.Size
-	a.Kind = jsonRequest.Kind
-	a.ContentType = jsonRequest.ContentType
-	a.Metadata = jsonRequest.Metadata
-
-	verification, err := user.Sign(a, pubKey, os.Getenv(meta.KeyStorePasswordEnv), state, parseVisibility(jsonRequest.Visibility))
-	if err != nil {
-		writeErrorResponse(w, "sign error", err, http.StatusBadRequest)
-		return
-	}
-
-	writeResponse(w, types.NewResult(&a, nil, verification))
-}
-
-func verify(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	hash := vars["hash"]
-	user, _ := currentUser()
-
-	verification, err := api.BlockChainVerify(hash)
-	if err != nil {
-		panic(err)
-	}
-
-	var artifact *api.ArtifactResponse
-	if !verification.Unknown() {
-		artifact, _ = api.LoadArtifactForHash(user, hash, verification.MetaHash())
-	}
-
-	writeResponse(w, types.NewResult(nil, artifact, verification))
-}
-
 func index(w http.ResponseWriter, r *http.Request) {
 	// can be used for healthcheck
 	fmt.Fprintln(w, "OK")
-}
-
-func parseVisibility(value string) meta.Visibility {
-	switch value {
-	case "0":
-		return meta.VisibilityPublic
-	case "1":
-		return meta.VisibilityPrivate
-	case "PUBLIC":
-		return meta.VisibilityPublic
-	case "PRIVATE":
-		return meta.VisibilityPrivate
-	default:
-		return meta.VisibilityPrivate
-	}
 }
